Refactored misassembly code to separate files. Fixed tests. Added missing dependencies.
