+++ requirements-dev.txt
@@ -1,7 +1,7 @@
 matplotlib==3.8.3
 numpy==1.26.4
 polars==1.6.0
-portion==2.4.2
+intervaltree==3.1.0
 pysam==0.22.0
 scipy==1.12.0
 pre-commit
