--- test/misjoin/config_static.toml
@@ -1,60 +1,0 @@
-[first]
-# Min horizontal distance between peaks.
-thr_min_peak_horizontal_distance = 1
-
-# Min width of peak to consider.
-thr_min_peak_width = 20
-
-# Min horizontal distance between valleys.
-thr_min_valley_horizontal_distance = 1
-
-# Min width of valley to consider.
-thr_min_valley_width = 10
-
-# Number of std above mean to include peak.
-thr_peak_height_std_above = 4
-
-# Number of std below mean to include valley.
-thr_valley_height_std_below = 3
-
-# Valleys with coverage below this threshold are considered misjoins.
-# If float:
-# * ex. 0.1 => Valley where median is less than or equal to 10% of mean
-# If int:
-# * ex. 2 => Valleys with median below 2
-thr_misjoin_valley = 3
-
-# Peaks with coverage above this threshold are considered collapses w/no variants.
-# If float:
-# * ex. 3.0 => Peaks where minimum is greater than or equal to 300% of mean
-# If int:
-# * ex. 40 => Peaks with minimum above 40
-thr_collapse_peak = 2.0
-
-# Group consecutive positions allowing a maximum gap of x.
-# Larger value groups more positions.
-valley_group_distance = 500
-
-peak_group_distance = 500
-
-[second]
-# Percent threshold of most freq base to allow second most freq base
-# 30 * 0.1 = 3 so above 3 is allowed.
-thr_min_perc_first = 0.1
-
-# Number of std above mean to include peak.
-thr_peak_height_std_above = 3
-
-# Group consecutive positions allowing a maximum gap of x.
-# Larger value groups more positions.
-group_distance = 30_000
-
-# Min group size.
-thr_min_group_size = 5
-
-[gaps]
-thr_max_allowed_gap_size = 1000
-
-[general]
-# Window size. Only applicable if no region bed file is provided.
-window_size = 5_000_000
