+++ test/misjoin/expected_perc.bed
@@ -1,1 +1,2 @@
-HG00171_chr16_haplotype1-0000003:1881763-8120526	3977787	3994078	HET
+HG00171_chr16_haplotype1-0000003:1881763-8120526	3958553	3994078	ERROR
+HG00171_chr16_haplotype1-0000003:1881763-8120526	5830987	5852649	MISJOIN
