--- test/misjoin/expected_static.bed
@@ -1,2 +1,0 @@
-HG00171_chr16_haplotype1-0000003:1881763-8120526	3977787	3994078	HET
-HG00171_chr16_haplotype1-0000003:1881763-8120526	5830987	5852649	MISJOIN
