+++ test/overlay/.gitignore
@@ -1,1 +1,2 @@
 output*/
+overlap_partial
