+++ test/standard/expected.bed
@@ -1,7 +1,8 @@
-haplotype2-0000133:3021508-8691473	3643515	3715420	HET
-haplotype2-0000133:3021508-8691473	5073878	5105033	HET
+haplotype2-0000133:3021508-8691473	3643515	3660435	HET
+haplotype2-0000133:3021508-8691473	3673449	3727518	ERROR
+haplotype2-0000133:3021508-8691473	5073878	5105033	ERROR
 haplotype2-0000133:3021508-8691473	5640379	5645142	HET
-haplotype2-0000133:3021508-8691473	5965734	5975145	HET
+haplotype2-0000133:3021508-8691473	5965734	5975145	ERROR
 haplotype2-0000133:3021508-8691473	7029969	7146182	GAP
 haplotype2-0000133:3021508-8691473	7570612	7586703	HET
 haplotype2-0000133:3021508-8691473	8388704	8398887	HET
